Design, Develop, and Deploy Multi-Agent Systems with CrewAI/M2/assignment/utils.py
# Add your utilities or helper functions to this file.
import os
from dotenv import load_dotenv, find_dotenv
import re

# these expect to find a .env file at the directory above the lesson.                                                                                                                     # the format for that file is (without the comment)                                                                                                                                       #API_KEYNAME=AStringThatIsTheLongAPIKeyFromSomeService
def load_env():
    _ = load_dotenv(find_dotenv())

def get_openai_api_key():
    load_env()
    openai_api_key = os.getenv("OPENAI_API_KEY")
    return openai_api_key

def get_serper_api_key():
    load_env()
    serper_api_key = os.getenv("SERPER_API_KEY")
    return serper_api_key

def clean_markdown(text):
    """Remove code fence markers if present"""
    text = re.sub(r'^```(?:markdown|md)?\s*\n', '', text, flags=re.IGNORECASE)
    text = re.sub(r'\n```\s*$', '', text)
    return text.strip()