Design, Develop, and Deploy Multi-Agent Systems with CrewAI/requirements.txt
# Core CrewAI Framework
crewai[tools]>=1.3.0

# Environment Variables Management
python-dotenv>=1.0.0

# Data Serialization
dill>=0.3.8

# Data Validation
pydantic>=2.0.0

# CrewAI Tools - Search and Web Scraping
crewai-tools>=0.20.0

# Additional dependencies that may be needed
# Jupyter for notebook support
jupyter>=1.0.0
ipython>=8.0.0

# Standard library (already included in Python)
# - os
# - sys
# - itertools
# - json
# - typing

